add bitbox-datetime workspace crate
It moves the C-wrapped datetime stuff from bitbox02/src/lib.rs.
