+++ src/rust/bitbox-datetime/Cargo.toml
@@ -1,0 +1,11 @@
+# SPDX-License-Identifier: Apache-2.0
+
+[package]
+name = "bitbox-datetime"
+version = "0.1.0"
+authors = ["Shift Crypto AG <support@bitbox.swiss>"]
+edition = "2024"
+description = "Date/time helpers for BitBox"
+license = "Apache-2.0"
+
+[dependencies]
