+++ src/rust/bitbox02-rust/src/hww/api/backup.rs
@@ -70,7 +70,7 @@ pub async fn create(
     hal.ui()
         .confirm(&confirm::Params {
             title: "Is today?",
-            body: &bitbox02::format_datetime(timestamp, timezone_offset, true)
+            body: &bitbox_datetime::format_datetime(timestamp, timezone_offset, true)
                 .map_err(|_| Error::InvalidInput)?,
             ..Default::default()
         })
